refactor(apns): extract payload builders
